+++ src/components/pages/markdown/components/link.js
@@ -9,8 +9,6 @@ exports.MarkdownLink = void 0;
 
 var _react = _interopRequireDefault(require("react"));
 
-var _styledComponents = _interopRequireDefault(require("styled-components"));
-
 var MarkdownLink = function MarkdownLink(props) {
   return _react.default.createElement("a", {
     href: props.href,