+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Fix: Update `modifierPargaprap` props to `MarkdownPage` component.
 
 ## [1.18.0] - 2019-02-27