+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Update `paragraphModifier` props to `MarkdownPage` component.
+## [1.18.1] - 2019-02-28
+
+Fix:
+- Update `paragraphModifier` props to `MarkdownPage` component.
 
 ## [1.18.0] - 2019-02-27